Return final ContentRuleList file after compilation instead of temporary file https://bugs.webkit.org/show_bug.cgi?id=279902
Reviewed by Timothy Hatcher.

After compiling a ContentRuleList, a map of the temporary file was
returned, instead of the final file.
This shared a map to a read-write filehandle to the WebProcess, while
it does not need to write the file.

Instead, share a read-only map of the final file.

Drive-by removal of trailing spaces.

* Source/WebKit/UIProcess/API/APIContentRuleListStore.cpp:
(API::compiledToFile):

Canonical link: https://commits.webkit.org/283910@main