REGRESSION(283879@main): When automatically resize is enabled, PDF unexpectedly scrolls when window is resized https://bugs.webkit.org/show_bug.cgi?id=280018 rdar://136311543
Reviewed by Aditya Keerthi.

It's wrong for us to ignore anchoring info if the document layout update
is not a no-op, which was a behavior change we introduced in 283487@main.

Instead, we backtrack our behavior into always respecting anchoring info
if present, and then introduce a slight modification where we recompute
the anchoring info if the document layout changes entirely on an update,
as would be the case when unlocking some encrypted PDF documents.

This modification is required so that we don't restore to a stale anchor
after unlocking some problematic encrypted files, while also not
throwing away anchoring information on minor layout changes.

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/PDFDocumentLayout.h:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::updateLayout):

Canonical link: https://commits.webkit.org/283960@main