[Curl] Fixes handling when download fails. https://bugs.webkit.org/show_bug.cgi?id=279647
Reviewed by Don Olmstead.

A NetworkProcess will crash if the instance of the Download class
indicated by m_pendingDownloadID is not registered in the download
manager and the download fails. To avoid this, we will modify it
to check m_client in the same way as the SOUP port.

* LayoutTests/platform/win/TestExpectations:
* Source/WebKit/NetworkProcess/curl/NetworkDataTaskCurl.cpp:
(WebKit::NetworkDataTaskCurl::curlDidFailWithError):

Canonical link: https://commits.webkit.org/283993@main