[Subgrid] Implicitly named lines can be empty due to inherited named lines https://bugs.webkit.org/show_bug.cgi?id=280068 rdar://136174078
Reviewed by Alan Baradlay.

For subgrids when we try to resolve a grid-area to line names we will first generate the
implicit lines associated with that area. If we end up actually generating these implicit
line names we will then override these line names with the line names that we inherited
from the actual grid. During this process we not actually end up inheriting any lines for
the particular name we are looking for (the one specified by the grid-area property)
because that track may be beyond the bounds specified for the subgrid. As a result, this
could end up in an empty Vector for the implicitly named lines.

In the testcase the subgrid attempts to create two named columns but that does not work
because the subgrid is only contained within the first column of the parent grid. As a
result, the subgrid's item cannot resolve the names of the lines associated with the
grid-area.

* LayoutTests/fast/grid/subgrid-inherited-line-names-crash-expected.txt: Added.
* LayoutTests/fast/grid/subgrid-inherited-line-names-crash.html: Added.
* Source/WebCore/rendering/style/GridPositionsResolver.cpp:
(WebCore::NamedLineCollectionBase::hasNamedLines const):

Canonical link: https://commits.webkit.org/283998@main