Proactively cleanup nw_resolve requests in case they don't progress. https://bugs.webkit.org/show_bug.cgi?id=279997 rdar://133085594
Reviewed by Chris Dumez and Youenn Fablet.

We have noticed huge number of pending requests in some cases. They are
not progressing for some reason and this change is to do a proactive cleanup.
60 seconds should be enough to resolve or cancel.

It is possible that a change is also needed on the networking stack.

* Source/WebCore/platform/network/cf/DNSResolveQueueCFNet.cpp:
(WebCore::DNSResolveQueueCFNet::performDNSLookup):

Canonical link: https://commits.webkit.org/284012@main