Fix WebContentProcess crashes after 284006@main https://bugs.webkit.org/show_bug.cgi?id=280095 rdar://136391408
Reviewed by Geoffrey Garen.

Based on an issue found by the smart pointer static analyzer, to ensure
that the WebProcessPool is alive while making a call to addMessageReceiver()
in the WebGeolocationManagerProxy constructor, 284006@main added a
protectedProcessPool() function.

The problem is that we are ref'ing the C++ wrapper of the Objective-C process
pool while it's being constructed (The process pool's constructor is what
calls the constructor of WebGeolocationManagerProxy). This causes
WebContentProcess crashes, so we're undoing this change for now until we can
figure out a better way to make this code pointer safe.

* Source/WebKit/UIProcess/WebContextSupplement.h:
(WebKit::WebContextSupplement::protectedProcessPool): Deleted.
* Source/WebKit/UIProcess/WebGeolocationManagerProxy.cpp:
(WebKit::WebGeolocationManagerProxy::WebGeolocationManagerProxy):

Canonical link: https://commits.webkit.org/284023@main