Fix PLT with site isolation enabled https://bugs.webkit.org/show_bug.cgi?id=280100 rdar://136397294
Reviewed by Ryosuke Niwa.

PLT has a site in it that is loaded in the main frame and then loaded in
an iframe under a different site's main frame.  Because the WebProcessPool's
process cache is currently designed and optimized for PSON, we can't use
the process cache the same way with site isolation when looking for a
process for an iframe because otherwise the BrowsingContextGroup's process
accounting gets messed up and we end up calling WebPage::loadRequest
with a frame identifier that doesn't exist in that process and it returns
early then the test hangs.  At some point we will need to teach the
WebProcessPool's process cache about site isolation to get perf up, but
until then we need to run the test to completion.

* Source/WebKit/UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::processForRegistrableDomain):

Canonical link: https://commits.webkit.org/284029@main