[CMake] MiniBrowser fails to link in non-developer builds with USE_SKIA=ON https://bugs.webkit.org/show_bug.cgi?id=280124
Reviewed by NOBODY (OOPS!).

Make Skia use the bmalloc/fastMalloc allocator only when using inside
WebKit. In order to enforce this, apply the same technique as for
ImageDiff: add SkMemory_malloc.cpp to the list of MiniBrowser sources,
and the linker will skip pulling the symbols from libSkia.a which
reference WTF, as they have less precedence.

* Source/WebCore/platform/Skia.cmake: List Skia in WebCore_PRIVATE_LIBRARIES.
* Source/WebKit/Platform/Skia.cmake: List Skia in WebKit_PRIVATE_LIBRARIES.
* Tools/ImageDiff/Skia.cmake: Correct comment, Skia links against
  WTF::fastMalloc but not bmalloc at the moment.
* Tools/MiniBrowser/wpe/CMakeLists.txt: Add the malloc Skia allocator as
  part of the MiniBrowser build.