REGRESSION(283414@main): [WPE][GTK] File descriptor leak if process launch is canceled immediately https://bugs.webkit.org/show_bug.cgi?id=280073
Reviewed by Carlos Garcia Campos.

If the socket monitor callback is not called, the server socket is
leaked. Oops.

I am separately working on converting ProcessLauncher and
IPC::Connection to use UnixFileDescriptor, which will make errors like
this less likely. But that's a large change, and we should fix this
with a small change.

* Source/WebKit/UIProcess/Launcher/ProcessLauncher.cpp:
* Source/WebKit/UIProcess/Launcher/ProcessLauncher.h:
* Source/WebKit/UIProcess/Launcher/glib/ProcessLauncherGLib.cpp:
(WebKit::ProcessLauncher::platformDestroy):
(WebKit::ProcessLauncher::launchProcess):

Canonical link: https://commits.webkit.org/284159@main