[ iOS ] css3/filters/backdrop-filter-page-scale.html is a flaky image failure. https://bugs.webkit.org/show_bug.cgi?id=280260 <rdar://134269886>
Reviewed by Wenson Hsieh.

The main change here is extending the fix in 258521@main to apply to the X axis
as well.

This test applies a page scale (with no offset) in the WebProcess (via a
WebPageProxy).  When this reaches WkWebViewIOS in the following
RemoteLayerTreeTransaction, it applies the same zoomScale, and zooms towards the
middle in the X axis. This leaves the left 200px offscreen (which includes all
the test content).  updateVisibleContentRects is then sent back to the
WebProcess with the new scroll offset, which calls setPageScaleFactor a second
time, with the scroll offset.

This test regressed by making it more async, which gives time for this multiple
transaction scrolling to take place, and then the test content is sometimes
offscreen.

This also fixes the TestRunner setPageScaleFactor setter to wait until a
rendering update has been done with the new page scale, before allowing tests to
complete.

* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.mm:
(-[WKWebView _updateScrollViewForTransaction:]):
* Source/WebKit/UIProcess/WebPageProxyTesting.cpp:
(WebKit::WebPageProxyTesting::setPageScaleFactor):
* Source/WebKit/WebProcess/WebPage/WebPageTesting.cpp:
(WebKit::WebPageTesting::setPageScaleFactor):
* Source/WebKit/WebProcess/WebPage/WebPageTesting.h:
* Source/WebKit/WebProcess/WebPage/WebPageTesting.messages.in:

Canonical link: https://commits.webkit.org/284197@main