[UnifiedPDF] [iOS] Keyboard presented for password field should not show an accessory view https://bugs.webkit.org/show_bug.cgi?id=280313 rdar://136635433
Reviewed by Aditya Keerthi.

Currently, the keyboard presented on a password field in WKPDFView does
not have an accessory view. As such, let's opt-out of an accessory view
in WKContentView if the web view is displaying a PDF document.

Note that we can drop a big hammer like this because the only situation
where we would present a keyboard for PDF content is for the password
field.

* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView requiresAccessoryView]):

Canonical link: https://commits.webkit.org/284202@main