Consistent handling of out-of-flow children in MathML layout functions
https://bugs.webkit.org/show_bug.cgi?id=179739

Reviewed by Frédéric Wang.

According to the spec [1], out-of-flow children should not affect the rendering of MathML elements. Ensure that out-of-flow children are still laid out and rendered, but that layout/validity logic only concerns in-flow children.

WebKit does not follow the spec specifically here ([2]), causing floating children to be OOF:

> The float property does not create floating of elements whose parent's computed display value is block math or inline math, and does not take them out-of-flow.

Cases in the WPT test in [3] used to pass regardless, since layout functions for MathML elements did not consider flow. Because they now *do* consider flow, the float property is no longer ignored in layout (as it causes elements to be OOF) and cases in the WPT test fail.

See webkit.org/b/279978.

[1] https://w3c.github.io/mathml-core/
[2] https://w3c.github.io/mathml-core/#css-styling
[3] https://github.com/web-platform-tests/wpt/blob/4d67f35a4dcc96bc40de00102a150d925d9fca61/mathml/relations/css-styling/floats/not-floating-001.html

Canonical link: https://commits.webkit.org/284225@main