Debug assertion on https://www.browserbench.org/MotionMark1.3.1/developer.html https://bugs.webkit.org/show_bug.cgi?id=280214
Reviewed by Said Abou-Hallawa.

When using threaded CPU rendering on WPE, clicking the  'Run Benchmark'
button on MotionMark developer.html fires an assertion. The ControlPart
object, associated with the button, is constructed during DisplayList
replay on a secondary thread, but destructed in the render tree teardown
phase during layout() on the main thread -- this leads to an assertion
in debug builds, that the construction/destruction thread is not
matched. Fix that by making ControlPart ThreadSafeRefCounted.

Covered by existing tests on WPE, in CPU rendering mode.

* Source/WebCore/platform/graphics/SystemImage.h:
* Source/WebCore/platform/graphics/controls/ControlPart.h:

Canonical link: https://commits.webkit.org/284244@main