[JSC] Convert CheckStructure with String Structure etc. to Check(StringUse) https://bugs.webkit.org/show_bug.cgi?id=280327 <rdar://problem/136659524>
Reviewed by Yijia Huang.

We found that CheckStructure(String structure) is not effectively
removed when abstract value is saying it is SpecString. We should
convert it to Check(StringUse) so that we can remove this based on more
coarse grained type information in abstract value. Also this patch does
it for SymbolUse and HeapBigIntUse too.

* Source/JavaScriptCore/dfg/DFGAbstractInterpreterInlines.h:
(JSC::DFG::AbstractInterpreter<AbstractStateType>::executeEffects):
* Source/JavaScriptCore/dfg/DFGConstantFoldingPhase.cpp:
(JSC::DFG::ConstantFoldingPhase::foldConstants):

Canonical link: https://commits.webkit.org/284249@main