[WPE][GTK] Handle EINTR in recvmsg() in IPC socket on Linux https://bugs.webkit.org/show_bug.cgi?id=280366
Reviewed by Adrian Perez de Castro.

recvmsg() may fail with EINTR in case it is interrupted by a signal.
In the referred bug this happens due to a child process exiting (with a
zero exit code indicating success). This causes SIGCHLD signal which
just happens to occur during the time recvmsg was waiting for a message.
As recvmsg invocation has no handling for EINTR it will fail, bringing
down whole application with it.

Handling EINTR in a while loop is a canonical solution used in
many places across WebKit codebase already.

* Source/WebKit/Platform/IPC/unix/ConnectionUnix.cpp:
(IPC::readPIDFromPeer):

Canonical link: https://commits.webkit.org/284274@main