Fix PLT without injected bundle with site isolation off https://bugs.webkit.org/show_bug.cgi?id=280430 rdar://136778475
Reviewed by Ryosuke Niwa.

In the ProvisionalPageProxy constructor, when doing a cross-site navigation
we create a new WebFrameProxy to be the main frame of the page after the navigation
if the navigation commits.  When swapping the WebPageProxy's main frame, we need
to add the PageLoadTimingFrameLoadStateObserver as an observer of the new main
frame's state or PageLoadTimingFrameLoadStateObserver::didFinishLoad will never
be called.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::swapToProvisionalPage):

Canonical link: https://commits.webkit.org/284333@main