[editing] Add offset range check in CompositeEditCommand::replaceSelectedTextInNode().
https://bugs.webkit.org/show_bug.cgi?id=150969

Reviewed by Ryosuke Niwa and Darin Adler.

When replacing text in a text node, we need to valid the offset
range against the length of the text node. For example, in the
fixed test, the length of text node is 0, in which case, there is
no need to call replaceTextInNode() function.

* LayoutTests/TestExpectations:
* Source/WebCore/editing/InsertTextCommand.cpp:
(WebCore::InsertTextCommand::doApply):

Canonical link: https://commits.webkit.org/284343@main