[WPE] Auxiliary processes fail to launch when ProcessProviderLibWPE is enabled https://bugs.webkit.org/show_bug.cgi?id=280321
Reviewed by Michael Catanzaro.

The ProcessProviderLibWPE is only used in the UIProcess, but
AuxiliaryProcessMainCommon::parseCommandLine attempts to check if
ProcessProviderLibWPE is enabled. Since ProcessProviderLibWPE is never
enabled for auxiliary processes, this causes a mismatch in the argument count
for auxiliary processes in runtimes where ProcessProviderLibWPE is used.
As a result, the auxiliary process never starts.

* Source/WebKit/Shared/unix/AuxiliaryProcessMain.cpp:
(WebKit::AuxiliaryProcessMainCommon::parseCommandLine):

Canonical link: https://commits.webkit.org/284344@main