Run in cross-origin frame testharness.js test fixes broke some other test in cross-origin frames https://bugs.webkit.org/show_bug.cgi?id=280516 rdar://136831018
Reviewed by NOBODY (OOPS!).

Previous changes would add a url search query parameter to indicate
that the test is run in cross-origin frame mode. However, this parameter
would break tests that assumed the query be empty.
Change the detection to key on 127.1.1.1:8080/root. The root alias
was added to support runInCrossOriginFrame.

* LayoutTests/fast/dom/empty-hash-and-search.html:
Update a test that would timeout with search query. Fix the error path
to work without an exception.
* LayoutTests/resources/testharnessreport.js:
* Tools/WebKitTestRunner/TestInvocation.cpp:
(WTR::TestInvocation::loadTestInCrossOriginIframe):
(WTR::addQueryParameter): Deleted.